# apps/api/internal/http/handlers/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enterprise File Platform API",
    "version": "0.1.0",
    "description": "Local-first ingestion orchestration and search API for file parsing, job execution, and operational monitoring.",
    "contact": {
      "name": "Project Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development API"
    }
  ],
  "paths": {
...
    "/api/v1/files": {
      "get": {
        "summary": "List files",
        "operationId": "listFiles",
        "tags": ["Files"],
        "parameters": [
          {"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "status", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "processing_status", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "extension", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "detected_file_type", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}, "example": 1},
          {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200}, "example": 25},
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200}, "description": "Deprecated alias for page_size"}
        ],
        "responses": {
          "200": {
            "description": "Paginated file list",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/FileListResponse"},
                "example": {
                  "total": 12,
                  "page": 1,
                  "page_size": 25,
                  "files": [
                    {
                      "id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                      "tenant_id": "11111111-1111-1111-1111-111111111001",
                      "parent_file_id": null,
                      "upload_id": "d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a",
                      "original_name": "sample.zip",
                      "normalized_name": "sample.zip",
                      "extension": ".zip",
                      "detected_mime_type": "application/zip",
                      "detected_file_type": "archive",
                      "storage_path": "uploads/11111111-1111-1111-1111-111111111001/d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a/sample.zip",
                      "size_bytes": 12345,
                      "sha256_hash": null,
                      "depth": 0,
                      "is_archive": true,
                      "is_password_protected": false,
                      "processing_status": "completed",
                      "created_at": "2026-06-16T11:40:12Z",
                      "updated_at": "2026-06-16T11:41:00Z"
                    }
                  ]
                }
              }
            }
          },
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
    "/api/v1/files/{file_id}/password": {
      "post": {
        "summary": "Submit password for password-protected archive",
        "operationId": "submitFilePassword",
        "tags": ["Files", "Security"],
        "parameters": [
          {"name": "file_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SubmitFilePasswordRequest"},
              "example": {"password": "change-me"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password submitted and job requeued",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SubmitFilePasswordResponse"},
                "example": {
                  "file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                  "job_id": "6d5fbf3a-21f8-4f2f-8bf7-cd6e9f13d9d6",
                  "status": "queued"
                }
              }
            }
          },
          "400": {
            "description": "Bad password request",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"error": "invalid file password request: password is required"}
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
    "/api/v1/files/{file_id}/children": {
      "get": {
        "summary": "List direct file children",
        "operationId": "listFileChildren",
        "tags": ["Files"],
        "parameters": [
          {"name": "file_id", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}, "example": 1},
          {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200}, "example": 25}
        ],
        "responses": {
          "200": {
            "description": "Paginated direct children",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/FileChildrenResponse"},
                "example": {
                  "file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                  "total": 3,
                  "page": 1,
                  "page_size": 25,
                  "children": []
                }
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
    "/api/v1/files/{file_id}/tree": {
      "get": {
        "summary": "Get recursive file tree rooted at a file",
        "operationId": "getFileTree",
        "tags": ["Files"],
        "parameters": [
          {"name": "file_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Nested file tree",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/FileTreeNode"},
                "example": {
                  "file": {
                    "id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                    "tenant_id": "11111111-1111-1111-1111-111111111001",
                    "original_name": "sample.zip",
                    "size_bytes": 12345,
                    "extension": ".zip",
                    "detected_file_type": "archive",
                    "storage_path": "uploads/11111111-1111-1111-1111-111111111001/d3ef1ac4-2f4f-4d8d-95d5-a7f3f5a8eb4a/sample.zip",
                    "depth": 0,
                    "is_archive": true,
                    "processing_status": "completed",
                    "created_at": "2026-06-16T11:40:12Z",
                    "updated_at": "2026-06-16T11:41:00Z"
                  },
                  "children": []
                }
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
    "/api/v1/files/{file_id}/records": {
      "get": {
        "summary": "List parsed records for a file",
        "operationId": "listFileRecords",
        "tags": ["Files", "Parsed Records"],
        "parameters": [
          {"name": "file_id", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1}, "example": 1},
          {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 200}, "example": 25}
        ],
        "responses": {
          "200": {
            "description": "Paginated parsed records",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/FileRecordsResponse"},
                "example": {
                  "file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                  "total": 10,
                  "page": 1,
                  "page_size": 25,
                  "records": [
                    {
                      "id": "9f5b4f0f-f7d8-4b9f-a1c4-2dd7d5a7d111",
                      "tenant_id": "11111111-1111-1111-1111-111111111001",
                      "file_id": "f6beef3c-8a45-4b89-8a3a-3d6d5b2d7f88",
                      "job_id": "6d5fbf3a-21f8-4f2f-8bf7-cd6e9f13d9d6",
                      "record_type": "text_line",
                      "line_number": 1,
                      "content_text": "user=john@example.com ip=198.51.100.12",
                      "structured_data": {"raw": "user=john@example.com ip=198.51.100.12"},
                      "extracted_entities": {"emails": ["john@example.com"], "ip_addresses": ["198.51.100.12"]},
                      "created_at": "2026-06-16T11:41:20Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "invalid request"}
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "resource not found"}
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "internal server error"}
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string"}
        },
        "required": ["error"]
      },
...
      "File": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string"},
          "parent_file_id": {"type": "string", "format": "uuid", "nullable": true},
          "upload_id": {"type": "string", "format": "uuid", "nullable": true},
          "original_name": {"type": "string"},
          "normalized_name": {"type": "string", "nullable": true},
          "extension": {"type": "string", "nullable": true},
          "detected_mime_type": {"type": "string", "nullable": true},
          "detected_file_type": {"type": "string", "nullable": true},
          "storage_path": {"type": "string"},
          "size_bytes": {"type": "integer", "format": "int64"},
          "sha256_hash": {"type": "string", "nullable": true},
          "depth": {"type": "integer"},
          "is_archive": {"type": "boolean"},
          "is_password_protected": {"type": "boolean"},
          "processing_status": {"type": "string"},
          "created_by": {"type": "string", "nullable": true},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "FileListResponse": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer"},
          "page_size": {"type": "integer"},
          "files": {"type": "array", "items": {"$ref": "#/components/schemas/File"}}
        }
      },
      "FileChildrenResponse": {
        "type": "object",
        "properties": {
          "file_id": {"type": "string", "format": "uuid"},
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer"},
          "page_size": {"type": "integer"},
          "children": {"type": "array", "items": {"$ref": "#/components/schemas/File"}}
        }
      },
      "ParsedRecord": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string"},
          "file_id": {"type": "string", "format": "uuid"},
          "job_id": {"type": "string", "format": "uuid"},
          "record_type": {"type": "string", "nullable": true},
          "record_number": {"type": "integer", "format": "int64", "nullable": true},
          "line_number": {"type": "integer", "format": "int64", "nullable": true},
          "chunk_number": {"type": "integer", "format": "int64", "nullable": true},
          "start_line": {"type": "integer", "format": "int64", "nullable": true},
          "end_line": {"type": "integer", "format": "int64", "nullable": true},
          "content_text": {"type": "string", "nullable": true},
          "structured_data": {"type": "object", "additionalProperties": true},
          "extracted_entities": {"type": "object", "additionalProperties": true},
          "event_timestamp": {"type": "string", "format": "date-time", "nullable": true},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "FileRecordsResponse": {
        "type": "object",
        "properties": {
          "file_id": {"type": "string", "format": "uuid"},
          "total": {"type": "integer", "format": "int64"},
          "page": {"type": "integer"},
          "page_size": {"type": "integer"},
          "records": {"type": "array", "items": {"$ref": "#/components/schemas/ParsedRecord"}}
        }
      },
      "FileTreeNode": {
        "type": "object",
        "properties": {
          "file": {"$ref": "#/components/schemas/File"},
          "children": {"type": "array", "items": {"$ref": "#/components/schemas/FileTreeNode"}}
        }
      },
      "SubmitFilePasswordRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": {"type": "string", "format": "password"}
        }
      },
      "SubmitFilePasswordResponse": {
        "type": "object",
        "properties": {
          "file_id": {"type": "string", "format": "uuid"},
          "job_id": {"type": "string", "format": "uuid"},
          "status": {"type": "string"}
        }
      },
...
    }
  },
...
}